+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 #### Bug Fixes
 
+- Fixed a bug that `TRY_CAST` reader option is ignored when calling `DataFrameReader.schema().csv()`.
+
 #### Improvements
 
 - Restored the following query improvements that were reverted in 1.47.0 due to bugs: